requirements.txt
bcrypt
certifi
cffi
click
Flask
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
Pillow
pycparser
six
SQLAlchemy
Werkzeug
Pillow
WTForms
Flask-Mail
export EMAIL_USER=renatas.gorodeckas2
export EMAIL_PASSWOR=<password>
